Fix potentential arithmetic overflows detected by msvc static analysis
git-svn-id: svn://scribus.net/trunk/Scribus@25386 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/ui/hruler.h
@@ -77,7 +77,7 @@ public:
 	double offset() const { return m_offset; }
 	
 private:
-	int findRulerHandle(QPoint mp, double grabRadius);
+	int findRulerHandle(QPoint mp, int grabRadius);
 	
 	void paintEvent(QPaintEvent *e) override;
 	void mousePressEvent(QMouseEvent *m) override;
